move seer similarity call into separate module
